fix: Update all function signatures and structs from i64 to i32
Continue fixing SQLite type compatibility:
- auth/mod.rs: create_otp user_id parameter
- studio/service.rs: create_session user_id parameter
- routes/admin.rs: AdminInfo.user_id field
- routes/studio.rs: StudioAuth.user_id field
- routes/jobs.rs: query tuple types and route parameters for chats/activity

+++ server/src/auth/mod.rs
@@ -58,7 +58,7 @@ impl AuthService {
     // Create OTP code for user
     pub async fn create_otp(
         pool: &SqlitePool,
-        user_id: i64,
+        user_id: i32,
         code: &str,
     ) -> Result<OtpCode, sqlx::Error> {
         // Invalidate all previous unused codes
